\name{fBN-package}
\alias{fBN-package}
\alias{fBN}
\docType{package}
\title{
\packageTitle{fBN}
}
\description{
\packageDescription{fBN}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fBN}
\packageIndices{fBN}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fBN}

Maintainer: \packageMaintainer{fBN}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
