\name{FISHprobes}
\docType{data}
\alias{FISHprobes}
\title{Sample data for FBN package}
\description{
Four FISH probes information for the two sample Human Myeloma Cell Lines (AMO1 and NCI-H929).}
\usage{}
\format{tab delimited .txt file}
\details{}
\examples{}
\source{}
\references{
Agnelli L et al. (2009), "A SNP Microarray and FISH-Based Procedure to Detect Allelic Imbalances in Multiple Myeloma: an Integrated Genomics Approach Reveals a Wide Gene Dosage Effect", \emph{Genes Chrom Cancer}
}
\keyword{datasets}
