% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fBN_wrapper.R
\name{get_initial_parameter}
\alias{get_initial_parameter}
\title{Title}
\usage{
get_initial_parameter(
  xobs,
  zobs,
  time,
  numk = NULL,
  nknots = 20,
  tolCPV = 0.9,
  useAT = FALSE
)
}
\arguments{
\item{xobs}{a \enq{n × |tobs|} matrix where n is the number of samples. Its (i, j)the element indicates the discrete observation for functional covariate k at time point \enq{t^{h}_k} of sample i if \enq{j = \sum_{l = 1}^{k - 1}|t_l| + h} according to the arrangement of tobs. Missing values are allowed and specified as NaN.}

\item{zobs}{a \enq{n × q} matrix where q is the number of baseline scalar covariates}

\item{useAT}{}
}
\value{

}
\description{
Title
}
