\name{hmcls}
\docType{data}
\alias{hmcls}
\title{Sample data for FBN package}
\description{
Two Human Myeloma Cell Lines (AMO1 and NCI-H929) profiled on GeneChip Human Mapping 250K NspI arrays.}
\format{image data .rda}
\references{
Agnelli L et al. (2009), "A SNP Microarray and FISH-Based Procedure to Detect Allelic Imbalances in Multiple Myeloma: an Integrated Genomics Approach Reveals a Wide Gene Dosage Effect", \emph{Genes Chrom Cancer}
}
\keyword{datasets}
