\name{FBN-package}
\alias{FBN-package}
\alias{FBN}
\docType{package}
\title{
	FISH Based Normalization and Copy Number inference of SNP microarray data
}
\description{
  Normalizes the data from a file containing the raw values of the SNP probes of microarrray data
  by using the FISH probes and their corresponding CNs.
}
\details{
\tabular{ll}{
Package: \tab FBN\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-04-03\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
To start using the FBN package, call \code{library(FBN)}.
}
\author{ Adrian Andronache \email{adi.andronache@gmail.com} \cr
Luca Agnelli \email{luca.agnelli@gmail.com}
}
\references{
Agnelli L et al. (2009), "A SNP Microarray and FISH-Based Procedure to Detect Allelic Imbalances in Multiple Myeloma: an Integrated Genomics Approach Reveals a Wide Gene Dosage Effect", \emph{Genes Chrom Cancer}
}
\keyword{ package }

